requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
firebase-admin==6.5.0
gunicorn==22.0.0
requests==2.31.0
